Move LocalDialog to design components
